+++ src/main/java/com/example/shopping/common/util/JwtUtil.java
@@ -4,6 +4,8 @@ package com.example.shopping.common.util;
 import com.example.shopping.common.exception.ServerException;
 import com.example.shopping.domain.auth.entity.RefreshToken;
 import com.example.shopping.domain.auth.repository.RefreshTokenRepository;
+import com.example.shopping.domain.user.entity.User;
+import com.example.shopping.domain.user.repository.UserRepository;
 import com.example.shopping.domain.user.role.UserRole;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
@@ -13,19 +15,25 @@ import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
-
+import org.springframework.web.server.ResponseStatusException;
 
 
 import java.security.Key;
 import java.util.Base64;
 import java.util.Date;
 
+import static com.example.shopping.common.exception.ErrorCode.INVALID_TOKEN;
+import static com.example.shopping.common.exception.ErrorCode.USER_NOT_FOUND;
+
 @Component
 @RequiredArgsConstructor
 public class JwtUtil {
 
     private final RefreshTokenRepository refreshTokenRepository;
+    private final UserRepository userRepository;
+    private final JwtUtil jwtUtil;
 
     private static final String BEARER_PREFIX = "Bearer ";
     private static final long TOKEN_TIME = 60 * 60 * 1000L; //60분
@@ -89,5 +97,31 @@ public class JwtUtil {
                 .build()
                 .parseClaimsJws(token)
                 .getBody();
+    }
+
+    @Transactional(readOnly = true)
+    public RefreshToken getRefreshToken(String accessToken) {
+        RefreshToken refreshToken = refreshTokenRepository.findByAccessToken(accessToken).orElseThrow(
+                ()-> new ResponseStatusException(INVALID_TOKEN.getStatus(),INVALID_TOKEN.getMessage()));
+        return refreshToken;
+    }
+
+    @Transactional
+    public void removeRefreshToken(String accessToken) {
+        refreshTokenRepository.findByAccessToken(accessToken)
+                .ifPresent(refreshTokenRepository::delete);
+    }
+
+    @Transactional
+    public String reCreateAccessToken(String originAccessToken, RefreshToken refreshToken) {
+        Long userId = refreshToken.getId();
+        User user = userRepository.findById(userId).orElseThrow(
+                ()-> new ResponseStatusException(USER_NOT_FOUND.getStatus(),USER_NOT_FOUND.getMessage()));
+        String newAccessToken = jwtUtil.createAccessToken(user.getId(),user.getEmail(),user.getRole(),user.getName(),user.getAddress());
+
+        removeRefreshToken(originAccessToken);
+        refreshTokenRepository.save(new RefreshToken(userId,newAccessToken,refreshToken.getRefreshToken()));
+        return newAccessToken;
+
     }
 }
